Move twitch components to TwitchContext
