+++ src/Components/Navbar.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { ButtonGroup, Button, HStack, Box, Flex, Heading, Spacer, Tooltip, IconButton, Menu, MenuButton, MenuItem, MenuList } from "@chakra-ui/react"
+import { ButtonGroup, Button, HStack, Box, Flex, Heading, Spacer, IconButton, Menu, MenuButton, MenuItem, MenuList } from "@chakra-ui/react"
 import { HamburgerIcon } from "@chakra-ui/icons";
